RAG-Anything/requirements.txt
huggingface_hub
# LightRAG packages
lightrag-hku
# MinerU 2.0 packages (replaces magic-pdf)
mineru[core]
# Progress bars for batch processing
tqdm
# Custom model dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
requests>=2.25.0
# Note: Optional dependencies are now defined in setup.py extras_require:
# - [image]: Pillow>=10.0.0 (for BMP, TIFF, GIF, WebP format conversion)
# - [text]: reportlab>=4.0.0 (for TXT, MD to PDF conversion)
# - [office]: requires LibreOffice (external program, not Python package)
# - [all]: includes all optional dependencies
#
# Install with: pip install raganything[image,text] or pip install raganything[all]
